LPDC-998: Rename utils.ts to rdf-utils.ts
