+++ rdf-utils.ts
